#369 Fix BC break introduced in 3.5
The BC break was caused by the changes to `ContentSecurityPolicyListener::getReport` and `ContentSecurityPolicyListener::getEnforcement`. In 3.5, these methods return a new DirectiveSet instead of the same instance as in previous versions. Code that used these methods to modify the DirectiveSets no longer saw these modifications reflected in the CSP header.

This is fixed by making `ConfigurationDirectiveSetBuilder` (the default builder) always return the same DirectiveSet by building it once rather than on each call.